moved images into dedicated directory
