+++ apps/web/src/components/WhyBase/EmpoweredByCoinbase.tsx
@@ -1,10 +1,8 @@
+import Image from 'apps/web/node_modules/next/image';
+
 import { ButtonWithLinkAndEventLogging } from 'apps/web/src/components/Button/ButtonWithLinkAndEventLogging';
 
-import empoweredByCoinbase from './EmpoweredByCoinbase.png';
-
-const imageBackgroundStyles = {
-  backgroundImage: `url('${empoweredByCoinbase.src}')`,
-};
+import empoweredByCoinbase from './images/EmpoweredByCoinbase.png';
 
 export default async function EmpoweredByCoinbase() {
   return (
@@ -17,10 +15,9 @@ export default async function EmpoweredByCoinbase() {
           <span>5.</span>
           <span className="ml-4">Empowered by Coinbase</span>
         </h2>
-        <div
-          className="h-[235px] w-[300px] self-center bg-contain bg-center bg-no-repeat sm:hidden"
-          style={imageBackgroundStyles}
-        />
+        <div className="h-[235px] w-[300px] self-center bg-contain bg-center bg-no-repeat sm:hidden">
+          <Image src={empoweredByCoinbase} alt="Empowered by Coinbase" />
+        </div>
         <div>
           <span className="text-base sm:text-lg">
             {`Base is incubated within Coinbase and plans to progressively decentralize in the years
@@ -41,10 +38,9 @@ export default async function EmpoweredByCoinbase() {
           </div>
         </div>
       </div>
-      <div
-        className="hidden h-auto w-full  bg-contain bg-center bg-no-repeat sm:block"
-        style={imageBackgroundStyles}
-      />
+      <div className="hidden h-auto w-full  bg-contain bg-center bg-no-repeat sm:block">
+        <Image src={empoweredByCoinbase} alt="Empowered by Coinbase" />
+      </div>
     </div>
   );
 }
