+++ apps/web/src/components/WhyBase/images/EmpoweredByCoinbase.png
