+++ apps/web/src/components/WhyBase/images/why-base-hero-bg.png
