+++ apps/web/src/components/WhyBase/partners.ts
@@ -1,27 +1,27 @@
 import { StaticImport } from 'apps/web/node_modules/next/dist/shared/lib/get-img-props';
 
-import zeroEx from 'apps/web/src/components/WhyBase/partnerImages/0x.svg';
-import aave from 'apps/web/src/components/WhyBase/partnerImages/aave.svg';
-import animoca from 'apps/web/src/components/WhyBase/partnerImages/animoca.svg';
-import blockdaemon from 'apps/web/src/components/WhyBase/partnerImages/blockdaemon.svg';
-import blockscout from 'apps/web/src/components/WhyBase/partnerImages/blockscout.svg';
-import chainlink from 'apps/web/src/components/WhyBase/partnerImages/chainlink.svg';
-import daz3d from 'apps/web/src/components/WhyBase/partnerImages/daz3d.svg';
-import dune from 'apps/web/src/components/WhyBase/partnerImages/dune.svg';
-import etherscan from 'apps/web/src/components/WhyBase/partnerImages/etherscan.svg';
-import euler from 'apps/web/src/components/WhyBase/partnerImages/euler.svg';
-import flipside from 'apps/web/src/components/WhyBase/partnerImages/flipside.svg';
-import hop from 'apps/web/src/components/WhyBase/partnerImages/hop.svg';
-import infura from 'apps/web/src/components/WhyBase/partnerImages/infura.svg';
-import magiceden from 'apps/web/src/components/WhyBase/partnerImages/magiceden.svg';
-import mnemonic from 'apps/web/src/components/WhyBase/partnerImages/mnemonic.svg';
-import nansen from 'apps/web/src/components/WhyBase/partnerImages/nansen.svg';
-import parallel from 'apps/web/src/components/WhyBase/partnerImages/parallel.svg';
-import quicknode from 'apps/web/src/components/WhyBase/partnerImages/quicknode.svg';
-import socket from 'apps/web/src/components/WhyBase/partnerImages/socket.svg';
-import sushiswap from 'apps/web/src/components/WhyBase/partnerImages/sushiswap.svg';
-import thirdweb from 'apps/web/src/components/WhyBase/partnerImages/thirdweb.svg';
-import transak from 'apps/web/src/components/WhyBase/partnerImages/transak.svg';
+import zeroEx from 'apps/web/src/components/WhyBase/images/partnerImages/0x.svg';
+import aave from 'apps/web/src/components/WhyBase/images/partnerImages/aave.svg';
+import animoca from 'apps/web/src/components/WhyBase/images/partnerImages/animoca.svg';
+import blockdaemon from 'apps/web/src/components/WhyBase/images/partnerImages/blockdaemon.svg';
+import blockscout from 'apps/web/src/components/WhyBase/images/partnerImages/blockscout.svg';
+import chainlink from 'apps/web/src/components/WhyBase/images/partnerImages/chainlink.svg';
+import daz3d from 'apps/web/src/components/WhyBase/images/partnerImages/daz3d.svg';
+import dune from 'apps/web/src/components/WhyBase/images/partnerImages/dune.svg';
+import etherscan from 'apps/web/src/components/WhyBase/images/partnerImages/etherscan.svg';
+import euler from 'apps/web/src/components/WhyBase/images/partnerImages/euler.svg';
+import flipside from 'apps/web/src/components/WhyBase/images/partnerImages/flipside.svg';
+import hop from 'apps/web/src/components/WhyBase/images/partnerImages/hop.svg';
+import infura from 'apps/web/src/components/WhyBase/images/partnerImages/infura.svg';
+import magiceden from 'apps/web/src/components/WhyBase/images/partnerImages/magiceden.svg';
+import mnemonic from 'apps/web/src/components/WhyBase/images/partnerImages/mnemonic.svg';
+import nansen from 'apps/web/src/components/WhyBase/images/partnerImages/nansen.svg';
+import parallel from 'apps/web/src/components/WhyBase/images/partnerImages/parallel.svg';
+import quicknode from 'apps/web/src/components/WhyBase/images/partnerImages/quicknode.svg';
+import socket from 'apps/web/src/components/WhyBase/images/partnerImages/socket.svg';
+import sushiswap from 'apps/web/src/components/WhyBase/images/partnerImages/sushiswap.svg';
+import thirdweb from 'apps/web/src/components/WhyBase/images/partnerImages/thirdweb.svg';
+import transak from 'apps/web/src/components/WhyBase/images/partnerImages/transak.svg';
 
 export const partners: Partner[] = [
   { name: 'transak', img: transak as StaticImport },
